Added OIDC discovery feature
When using annotated service, some annotations can make that get the OAS
spec from the upstream API. When these annotations are in place, the
controller will make an HTTP request to the defined endpoint and
retrieve the config.

Regarding testing, I went to unit Test instead of integration test. The
main reason is that the controller needs to be inside the Kubernetes
cluster to be able to use the internal service, so integration testing,
as it is, is complicated.

Unit testing is based on Golang Unit testing -no framework at all- soon;
maybe we should think about moving this to Ginkgo or anything similar to
this.

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>